SourceArea: Merge SourcePosition & -Range modules
This allows for mutual comparisons without depending on circular dependencies.
